Finished the test file - has to be run on a different machine, since tests dont seem to work for me.
